backend/requirements.txt
django>=4.2,<5
djangorestframework>=3.14
django-cors-headers>=4.3

